refactor: cleanup duplication, componentize
