+++ packages/core/__tests__/screens/TogglePushNotifications.test.tsx
@@ -1,19 +1,16 @@
 import { fireEvent, render, waitFor } from '@testing-library/react-native'
 import React from 'react'
 
-import { StoreProvider, defaultState } from '../../src/contexts/store'
-import { testIdWithKey } from '../../src/utils/testable'
-import PushNotification from '../../src/screens/PushNotification'
-import { CustomBasicAppContext } from '../helpers/app'
-import { MainContainer } from '../../src/container-impl'
 import { container } from 'tsyringe'
 import { TOKENS } from '../../src/container-api'
+import { MainContainer } from '../../src/container-impl'
+import { StoreProvider, defaultState } from '../../src/contexts/store'
+import TogglePushNotifications from '../../src/screens/TogglePushNotifications'
 import { Config } from '../../src/types/config'
+import { testIdWithKey } from '../../src/utils/testable'
+import { CustomBasicAppContext } from '../helpers/app'
 
-describe('PushNotification Screen', () => {
-  beforeEach(() => {
-    jest.clearAllMocks()
-  })
+describe('TogglePushNotification Screen', () => {
   const setup = jest.fn()
   const toggle = jest.fn()
   const status = async () => await Promise.resolve('denied')
@@ -22,11 +19,6 @@ describe('PushNotification Screen', () => {
   context.container.registerInstance(TOKENS.CONFIG, { ...config, enablePushNotifications: { status, setup, toggle } })
 
   test('Push notification screen renders correctly in settings', async () => {
-    const route = {
-      params: {
-        isMenu: true,
-      },
-    } as any
     const tree = render(
       <StoreProvider
         initialState={{
@@ -34,7 +26,7 @@ describe('PushNotification Screen', () => {
         }}
       >
         <CustomBasicAppContext container={context}>
-          <PushNotification route={route} navigation={jest.fn() as any} />
+          <TogglePushNotifications />
         </CustomBasicAppContext>
       </StoreProvider>
     )
@@ -47,28 +39,5 @@ describe('PushNotification Screen', () => {
     })
     expect(setup).toHaveBeenCalledTimes(1)
     expect(toggle).toHaveBeenCalledTimes(1)
-  })
-  test('Push notification screen renders correctly in onboarding', async () => {
-    const route = {
-      params: {},
-    } as any
-    const tree = render(
-      <StoreProvider
-        initialState={{
-          ...defaultState,
-        }}
-      >
-        <CustomBasicAppContext container={context}>
-          <PushNotification route={route} navigation={jest.fn() as any} />
-        </CustomBasicAppContext>
-      </StoreProvider>
-    )
-
-    const continueButton = tree.getByTestId(testIdWithKey('PushNotificationContinue'))
-    expect(continueButton).not.toBe(null)
-    await waitFor(() => {
-      fireEvent(continueButton, 'press')
-    })
-    expect(setup).toHaveBeenCalledTimes(1)
   })
 })
