+++ packages/core/src/screens/PINChange.tsx
@@ -1,6 +1,6 @@
 import { ParamListBase, useNavigation } from '@react-navigation/native'
 import { StackNavigationProp, StackScreenProps } from '@react-navigation/stack'
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import React, { useCallback, useMemo, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import {
   AccessibilityInfo,
@@ -12,14 +12,12 @@ import {
   TouchableOpacity,
   View,
 } from 'react-native'
-import Icon from 'react-native-vector-icons/MaterialIcons'
 
 // eslint-disable-next-line import/no-named-as-default
 import { ButtonType } from '../components/buttons/Button-api'
-import { InlineErrorType, InlineMessageProps } from '../components/inputs/InlineErrorText'
 import PINInput from '../components/inputs/PINInput'
+import PINValidationHelper from '../components/misc/PINValidationHelper'
 import AlertModal from '../components/modals/AlertModal'
-import { ThemedText } from '../components/texts/ThemedText'
 import KeyboardView from '../components/views/KeyboardView'
 import { EventTypes, minPINLength } from '../constants'
 import { TOKENS, useServices } from '../container-api'
@@ -27,19 +25,12 @@ import { useAnimatedComponents } from '../contexts/animated-components'
 import { useAuth } from '../contexts/auth'
 import { useStore } from '../contexts/store'
 import { useTheme } from '../contexts/theme'
+import { usePINValidation } from '../hooks/usePINValidation'
 import { HistoryCardType, HistoryRecord } from '../modules/history/types'
 import { BifoldError } from '../types/error'
 import { OnboardingStackParams, Screens } from '../types/navigators'
 import { useAppAgent } from '../utils/agent'
-import { createPINValidations, PINValidationsType } from '../utils/PINValidation'
 import { testIdWithKey } from '../utils/testable'
-
-interface ModalState {
-  visible: boolean
-  title: string
-  message: string
-  onModalDismiss?: () => void
-}
 
 const PINChange: React.FC<StackScreenProps<ParamListBase, Screens.ChangePIN>> = () => {
   const { agent } = useAppAgent()
@@ -48,24 +39,16 @@ const PINChange: React.FC<StackScreenProps<ParamListBase, Screens.ChangePIN>> = 
   const [PINTwo, setPINTwo] = useState('')
   const [PINOld, setPINOld] = useState('')
   const [isLoading, setIsLoading] = useState(false)
-  const [modalState, setModalState] = useState<ModalState>({
-    visible: false,
-    title: '',
-    message: '',
-  })
-  const iconSize = 24
   const navigation = useNavigation<StackNavigationProp<OnboardingStackParams>>()
   const [store] = useStore()
   const { t } = useTranslation()
-  const [inlineMessageField1, setInlineMessageField1] = useState<InlineMessageProps>()
-  const [inlineMessageField2, setInlineMessageField2] = useState<InlineMessageProps>()
   const { ColorPallet } = useTheme()
   const { ButtonLoading } = useAnimatedComponents()
   const PINTwoInputRef = useRef<TextInput>(null)
   const createPINButtonRef = useRef<TouchableOpacity>(null)
+
   const [
     PINHeader,
-    { PINSecurity },
     Button,
     inlineMessages,
     logger,
@@ -74,7 +57,6 @@ const PINChange: React.FC<StackScreenProps<ParamListBase, Screens.ChangePIN>> = 
     historyEventsLogger,
   ] = useServices([
     TOKENS.COMPONENT_PIN_HEADER,
-    TOKENS.CONFIG,
     TOKENS.COMP_BUTTON,
     TOKENS.INLINE_ERRORS,
     TOKENS.UTIL_LOGGER,
@@ -82,9 +64,17 @@ const PINChange: React.FC<StackScreenProps<ParamListBase, Screens.ChangePIN>> = 
     TOKENS.HISTORY_ENABLED,
     TOKENS.HISTORY_EVENTS_LOGGER,
   ])
-  const [PINOneValidations, setPINOneValidations] = useState<PINValidationsType[]>(
-    createPINValidations(PIN, PINSecurity.rules)
-  )
+
+  const {
+    PINValidations,
+    validatePINEntry,
+    inlineMessageField1,
+    inlineMessageField2,
+    modalState,
+    setModalState,
+    clearModal,
+    PINSecurity,
+  } = usePINValidation(PIN, PINTwo)
 
   const style = StyleSheet.create({
     screenContainer: {
@@ -99,51 +89,6 @@ const PINChange: React.FC<StackScreenProps<ParamListBase, Screens.ChangePIN>> = 
     controlsContainer: {},
   })
 
-  const attentionMessage = useCallback(
-    (title: string, message: string, pinOne: boolean) => {
-      if (inlineMessages.enabled) {
-        const config = {
-          message: message,
-          inlineType: InlineErrorType.error,
-          config: inlineMessages,
-        }
-        if (pinOne) {
-          setInlineMessageField1(config)
-        } else {
-          setInlineMessageField2(config)
-        }
-      } else {
-        setModalState({
-          visible: true,
-          title: title,
-          message: message,
-        })
-      }
-    },
-    [inlineMessages]
-  )
-
-  const validatePINEntry = useCallback(
-    (PINOne: string, PINTwo: string): boolean => {
-      for (const validation of PINOneValidations) {
-        if (validation.isInvalid) {
-          attentionMessage(
-            t('PINCreate.InvalidPIN'),
-            t(`PINCreate.Message.${validation.errorName}`, validation?.errorTextAddition),
-            true
-          )
-          return false
-        }
-      }
-      if (PINOne !== PINTwo) {
-        attentionMessage(t('PINCreate.InvalidPIN'), t('PINCreate.PINsDoNotMatch'), false)
-        return false
-      }
-      return true
-    },
-    [PINOneValidations, t, attentionMessage]
-  )
-
   const checkOldPIN = useCallback(
     async (PIN: string): Promise<boolean> => {
       const valid = await checkWalletPIN(PIN)
@@ -152,11 +97,12 @@ const PINChange: React.FC<StackScreenProps<ParamListBase, Screens.ChangePIN>> = 
           visible: true,
           title: t('PINCreate.InvalidPIN'),
           message: t(`PINCreate.Message.OldPINIncorrect`),
+          onModalDismiss: clearModal,
         })
       }
       return valid
     },
-    [checkWalletPIN, t]
+    [checkWalletPIN, t, setModalState, clearModal]
   )
 
   const logHistoryRecord = useCallback(() => {
@@ -200,6 +146,7 @@ const PINChange: React.FC<StackScreenProps<ParamListBase, Screens.ChangePIN>> = 
               message: t('PINChange.PinChangeSuccessMessage'),
               onModalDismiss: () => {
                 navigation.navigate(Screens.Settings as never)
+                clearModal()
               },
             })
           }
@@ -222,10 +169,6 @@ const PINChange: React.FC<StackScreenProps<ParamListBase, Screens.ChangePIN>> = 
     return PIN === '' || PINTwo === '' || PINOld === '' || PIN.length < minPINLength || PINTwo.length < minPINLength
   }, [inlineMessages, isLoading, PIN, PINTwo, PINOld])
 
-  useEffect(() => {
-    setInlineMessageField1(undefined)
-    setInlineMessageField2(undefined)
-  }, [PIN, PINTwo])
 
   return (
     <KeyboardView>
@@ -244,17 +187,11 @@ const PINChange: React.FC<StackScreenProps<ParamListBase, Screens.ChangePIN>> = 
             label={t('PINChange.EnterPINTitle')}
             onPINChanged={(p: string) => {
               setPIN(p)
-              setPINOneValidations(createPINValidations(p, PINSecurity.rules))
-
-              if (p.length === minPINLength) {
-                if (PINTwoInputRef?.current) {
-                  PINTwoInputRef.current.focus()
-                  // NOTE:(jl) `findNodeHandle` will be deprecated in React 18.
-                  // https://reactnative.dev/docs/new-architecture-library-intro#preparing-your-javascript-codebase-for-the-new-react-native-renderer-fabric
-                  const reactTag = findNodeHandle(PINTwoInputRef.current)
-                  if (reactTag) {
-                    AccessibilityInfo.setAccessibilityFocus(reactTag)
-                  }
+              if (p.length === minPINLength && PINTwoInputRef?.current) {
+                PINTwoInputRef.current.focus()
+                const reactTag = findNodeHandle(PINTwoInputRef.current)
+                if (reactTag) {
+                  AccessibilityInfo.setAccessibilityFocus(reactTag)
                 }
               }
             }}
@@ -269,13 +206,9 @@ const PINChange: React.FC<StackScreenProps<ParamListBase, Screens.ChangePIN>> = 
               setPINTwo(p)
               if (p.length === minPINLength) {
                 Keyboard.dismiss()
-                if (createPINButtonRef?.current) {
-                  // NOTE:(jl) `findNodeHandle` will be deprecated in React 18.
-                  // https://reactnative.dev/docs/new-architecture-library-intro#preparing-your-javascript-codebase-for-the-new-react-native-renderer-fabric
-                  const reactTag = findNodeHandle(createPINButtonRef.current)
-                  if (reactTag) {
-                    AccessibilityInfo.setAccessibilityFocus(reactTag)
-                  }
+                const reactTag = createPINButtonRef?.current && findNodeHandle(createPINButtonRef.current)
+                if (reactTag) {
+                  AccessibilityInfo.setAccessibilityFocus(reactTag)
                 }
               }
             }}
@@ -286,41 +219,13 @@ const PINChange: React.FC<StackScreenProps<ParamListBase, Screens.ChangePIN>> = 
             inlineMessage={inlineMessageField2}
           />
           {PINSecurity.displayHelper && (
-            <View style={{ marginBottom: 16 }}>
-              {PINOneValidations.map((validation, index) => {
-                return (
-                  <View style={{ flexDirection: 'row' }} key={index}>
-                    {validation.isInvalid ? (
-                      <Icon
-                        accessibilityLabel={t('PINCreate.Helper.ClearIcon')}
-                        name="clear"
-                        size={iconSize}
-                        color={ColorPallet.notification.errorIcon}
-                      />
-                    ) : (
-                      <Icon
-                        accessibilityLabel={t('PINCreate.Helper.CheckIcon')}
-                        name="check"
-                        size={iconSize}
-                        color={ColorPallet.notification.successIcon}
-                      />
-                    )}
-                    <ThemedText style={{ paddingLeft: 4 }}>
-                      {t(`PINCreate.Helper.${validation.errorName}`, validation?.errorTextAddition)}
-                    </ThemedText>
-                  </View>
-                )
-              })}
-            </View>
+            <PINValidationHelper validations={PINValidations} />
           )}
           {modalState.visible && (
             <AlertModal
               title={modalState.title}
               message={modalState.message}
-              submit={() => {
-                modalState.onModalDismiss?.()
-                setModalState({ ...modalState, visible: false, onModalDismiss: undefined })
-              }}
+              submit={modalState.onModalDismiss}
             />
           )}
         </View>
