+++ packages/core/src/screens/PINCreate.tsx
@@ -1,22 +1,22 @@
 import { ParamListBase } from '@react-navigation/native'
 import { StackScreenProps } from '@react-navigation/stack'
-import React, { useState, useRef, useEffect, useCallback, useMemo } from 'react'
+import React, { useCallback, useMemo, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import {
   AccessibilityInfo,
+  DeviceEventEmitter,
   Keyboard,
   StyleSheet,
-  View,
   TextInput,
   TouchableOpacity,
+  View,
   findNodeHandle,
-  DeviceEventEmitter,
 } from 'react-native'
-import Icon from 'react-native-vector-icons/MaterialIcons'
 
 // eslint-disable-next-line import/no-named-as-default
 import { ButtonType } from '../components/buttons/Button-api'
 import PINInput from '../components/inputs/PINInput'
+import PINValidationHelper from '../components/misc/PINValidationHelper'
 import AlertModal from '../components/modals/AlertModal'
 import KeyboardView from '../components/views/KeyboardView'
 import { EventTypes, minPINLength } from '../constants'
@@ -26,23 +26,14 @@ import { useAuth } from '../contexts/auth'
 import { DispatchAction } from '../contexts/reducers/store'
 import { useStore } from '../contexts/store'
 import { useTheme } from '../contexts/theme'
+import { usePINValidation } from '../hooks/usePINValidation'
 import { BifoldError } from '../types/error'
 import { Screens } from '../types/navigators'
-import { createPINValidations, PINValidationsType } from '../utils/PINValidation'
 import { testIdWithKey } from '../utils/testable'
-import { InlineErrorType, InlineMessageProps } from '../components/inputs/InlineErrorText'
-import { ThemedText } from '../components/texts/ThemedText'
 
 interface PINCreateProps extends StackScreenProps<ParamListBase, Screens.CreatePIN> {
   setAuthenticated: (status: boolean) => void
   explainedStatus: boolean
-}
-
-interface ModalState {
-  visible: boolean
-  title: string
-  message: string
-  onModalDismiss?: () => void
 }
 
 const PINCreate: React.FC<PINCreateProps> = ({ setAuthenticated, explainedStatus }) => {
@@ -50,22 +41,14 @@ const PINCreate: React.FC<PINCreateProps> = ({ setAuthenticated, explainedStatus
   const [PIN, setPIN] = useState('')
   const [PINTwo, setPINTwo] = useState('')
   const [isLoading, setIsLoading] = useState(false)
-  const [modalState, setModalState] = useState<ModalState>({
-    visible: false,
-    title: '',
-    message: '',
-  })
-  const iconSize = 24
   const [, dispatch] = useStore()
   const { t } = useTranslation()
-  const [inlineMessageField1, setInlineMessageField1] = useState<InlineMessageProps>()
-  const [inlineMessageField2, setInlineMessageField2] = useState<InlineMessageProps>()
 
   const { ColorPallet } = useTheme()
   const { ButtonLoading } = useAnimatedComponents()
   const PINTwoInputRef = useRef<TextInput>(null)
   const createPINButtonRef = useRef<TouchableOpacity>(null)
-  const [PINExplainer, PINHeader, { PINSecurity, showPINExplainer }, Button, inlineMessages] = useServices([
+  const [PINExplainer, PINHeader, { showPINExplainer }, Button, inlineMessages] = useServices([
     TOKENS.SCREEN_PIN_EXPLAINER,
     TOKENS.COMPONENT_PIN_HEADER,
     TOKENS.CONFIG,
@@ -75,9 +58,14 @@ const PINCreate: React.FC<PINCreateProps> = ({ setAuthenticated, explainedStatus
 
   const [explained, setExplained] = useState(explainedStatus || showPINExplainer === false)
 
-  const [PINOneValidations, setPINOneValidations] = useState<PINValidationsType[]>(
-    createPINValidations(PIN, PINSecurity.rules)
-  )
+  const {
+    PINValidations,
+    validatePINEntry,
+    inlineMessageField1,
+    inlineMessageField2,
+    modalState,
+    PINSecurity
+  } = usePINValidation(PIN, PINTwo)
 
   const style = StyleSheet.create({
     screenContainer: {
@@ -115,51 +103,6 @@ const PINCreate: React.FC<PINCreateProps> = ({ setAuthenticated, explainedStatus
     [setWalletPIN, setAuthenticated, dispatch, t]
   )
 
-  const attentionMessage = useCallback(
-    (title: string, message: string, pinOne: boolean) => {
-      if (inlineMessages.enabled) {
-        const config = {
-          message: message,
-          inlineType: InlineErrorType.error,
-          config: inlineMessages,
-        }
-        if (pinOne) {
-          setInlineMessageField1(config)
-        } else {
-          setInlineMessageField2(config)
-        }
-      } else {
-        setModalState({
-          visible: true,
-          title: title,
-          message: message,
-        })
-      }
-    },
-    [inlineMessages]
-  )
-
-  const validatePINEntry = useCallback(
-    (PINOne: string, PINTwo: string): boolean => {
-      for (const validation of PINOneValidations) {
-        if (validation.isInvalid) {
-          attentionMessage(
-            t('PINCreate.InvalidPIN'),
-            t(`PINCreate.Message.${validation.errorName}`, validation?.errorTextAddition),
-            true
-          )
-          return false
-        }
-      }
-      if (PINOne !== PINTwo) {
-        attentionMessage(t('PINCreate.InvalidPIN'), t('PINCreate.PINsDoNotMatch'), false)
-        return false
-      }
-      return true
-    },
-    [PINOneValidations, t, attentionMessage]
-  )
-
   const handleCreatePinTap = useCallback(async () => {
     setIsLoading(true)
     if (validatePINEntry(PIN, PINTwo)) {
@@ -179,11 +122,6 @@ const PINCreate: React.FC<PINCreateProps> = ({ setAuthenticated, explainedStatus
     setExplained(true)
   }, [])
 
-  useEffect(() => {
-    setInlineMessageField1(undefined)
-    setInlineMessageField2(undefined)
-  }, [PIN, PINTwo])
-
   return explained ? (
     <KeyboardView>
       <View style={style.screenContainer}>
@@ -193,17 +131,11 @@ const PINCreate: React.FC<PINCreateProps> = ({ setAuthenticated, explainedStatus
             label={t('PINCreate.EnterPINTitle')}
             onPINChanged={(p: string) => {
               setPIN(p)
-              setPINOneValidations(createPINValidations(p, PINSecurity.rules))
-
-              if (p.length === minPINLength) {
-                if (PINTwoInputRef?.current) {
-                  PINTwoInputRef.current.focus()
-                  // NOTE:(jl) `findNodeHandle` will be deprecated in React 18.
-                  // https://reactnative.dev/docs/new-architecture-library-intro#preparing-your-javascript-codebase-for-the-new-react-native-renderer-fabric
-                  const reactTag = findNodeHandle(PINTwoInputRef.current)
-                  if (reactTag) {
-                    AccessibilityInfo.setAccessibilityFocus(reactTag)
-                  }
+              if (p.length === minPINLength && PINTwoInputRef?.current) {
+                PINTwoInputRef.current.focus()
+                const reactTag = findNodeHandle(PINTwoInputRef.current)
+                if (reactTag) {
+                  AccessibilityInfo.setAccessibilityFocus(reactTag)
                 }
               }
             }}
@@ -218,13 +150,9 @@ const PINCreate: React.FC<PINCreateProps> = ({ setAuthenticated, explainedStatus
               setPINTwo(p)
               if (p.length === minPINLength) {
                 Keyboard.dismiss()
-                if (createPINButtonRef?.current) {
-                  // NOTE:(jl) `findNodeHandle` will be deprecated in React 18.
-                  // https://reactnative.dev/docs/new-architecture-library-intro#preparing-your-javascript-codebase-for-the-new-react-native-renderer-fabric
-                  const reactTag = findNodeHandle(createPINButtonRef.current)
-                  if (reactTag) {
-                    AccessibilityInfo.setAccessibilityFocus(reactTag)
-                  }
+                const reactTag = createPINButtonRef?.current && findNodeHandle(createPINButtonRef.current)
+                if (reactTag) {
+                  AccessibilityInfo.setAccessibilityFocus(reactTag)
                 }
               }
             }}
@@ -235,41 +163,13 @@ const PINCreate: React.FC<PINCreateProps> = ({ setAuthenticated, explainedStatus
             inlineMessage={inlineMessageField2}
           />
           {PINSecurity.displayHelper && (
-            <View style={{ marginBottom: 16 }}>
-              {PINOneValidations.map((validation, index) => {
-                return (
-                  <View style={{ flexDirection: 'row' }} key={index}>
-                    {validation.isInvalid ? (
-                      <Icon
-                        accessibilityLabel={t('PINCreate.Helper.ClearIcon')}
-                        name="clear"
-                        size={iconSize}
-                        color={ColorPallet.notification.errorIcon}
-                      />
-                    ) : (
-                      <Icon
-                        accessibilityLabel={t('PINCreate.Helper.CheckIcon')}
-                        name="check"
-                        size={iconSize}
-                        color={ColorPallet.notification.successIcon}
-                      />
-                    )}
-                    <ThemedText style={{ paddingLeft: 4 }}>
-                      {t(`PINCreate.Helper.${validation.errorName}`, validation?.errorTextAddition)}
-                    </ThemedText>
-                  </View>
-                )
-              })}
-            </View>
+            <PINValidationHelper validations={PINValidations} />
           )}
           {modalState.visible && (
             <AlertModal
               title={modalState.title}
               message={modalState.message}
-              submit={() => {
-                modalState.onModalDismiss?.()
-                setModalState({ ...modalState, visible: false, onModalDismiss: undefined })
-              }}
+              submit={modalState.onModalDismiss}
             />
           )}
         </View>
